requirements.txt
albumentations
glfw
keyboard
kneed
matplotlib
numpy
opencv-python
pillow
pycairo
# pyceres
PyGLM
PyGObject
PyOpenGL
pytorch-lightning
PyYAML
scikit-learn
scipy
tensorboard
termcolor
timm
torch
torch_cluster
torchinfo
tqdm